Pass prop isActiveTab onto each child view
Gives child views an easy way to update themselves each time they
are shown. This is possible to do without this modification, if a state
is kept in the parent of the tabview and wired to each child manually,
the point of this PR is to automate this and reduce that overhead.